refactor: re-export MediaItemPalette from media_items package
